+++ libs/ngx-contextmenu/src/lib/helper/evaluate.spec.ts
@@ -16,22 +16,26 @@ describe('#evaluateIfFunction', () => {
   it('should return the result of the evaluation of value if it is a function', () => {
     const item = { id: 'item' };
     const actualResult = true;
-    const value = jasmine
-      .createSpy('functionValue')
-      .and.returnValue(actualResult);
+    const testValue = jest.fn();
+    const value = (...args: unknown[]) => {
+      testValue(...args);
+      return actualResult;
+    };
     const result = evaluateIfFunction(value, item);
-    expect(value).toHaveBeenCalledWith(item);
+    expect(testValue).toHaveBeenCalledWith(item);
     expect(result).toBe(true);
   });
 
   it('should return the result of the evaluation of value if it is a function', () => {
     const item = { id: 'item' };
     const actualResult = false;
-    const value = jasmine
-      .createSpy('functionValue')
-      .and.returnValue(actualResult);
+    const testValue = jest.fn();
+    const value = (...args: unknown[]) => {
+      testValue(...args);
+      return actualResult;
+    };
     const result = evaluateIfFunction(value, item);
-    expect(value).toHaveBeenCalledWith(item);
+    expect(testValue).toHaveBeenCalledWith(item);
     expect(result).toBe(false);
   });
 });